+++ app/controllers/aeternitas/web_ui/application_controller.rb
@@ -9,6 +9,12 @@ module Aeternitas
         render template: 'aeternitas/web_ui/dashboard/error', status: status
       end
 
+      def set_timerange
+        @from = DateTime.parse(params.require(:from))
+        @to = DateTime.parse(params.require(:to))
+        @resolution = get_resolution(@from, @to)
+      end
+
       def get_resolution(from, to)
         case (to.to_time - from.to_time)
           when 0.hours..2.hours
