+++ lib/aeternitas/web_ui.rb
@@ -1,8 +1,5 @@
 require 'aeternitas/web_ui/engine'
 require 'aeternitas/web_ui/color_generator'
-require 'aeternitas/web_ui/dashboard_statistics'
-require 'aeternitas/web_ui/pollable_statistics'
-require 'aeternitas/web_ui/pollables_index_statistics'
 require 'aeternitas/web_ui/version'
 require 'jbuilder'
